refactor(models): split career models into separate files under internal/model/career
Move career model types from monolithic internal/repository/models/career.go
into individual files under internal/model/career/:
- string_slice.go: StringSlice type with SQL scanner/valuer
- skill.go: Skill model with GORM tags
- event.go: Event model with GORM tags
- fact.go: Fact model with GORM tags
- burst.go: Burst model with relationships

Also update pre-commit hook to skip testdata directories in golangci-lint
and test checks (they contain fake packages for analyzer testing).

Update noinlinecareer analyzer:
- Match model/career package path (in addition to domain/career)
- Exclude repository/career package (legitimate GORM model usage)

This improves code organisation and follows single-responsibility principle.
All imports updated across repository layer and analyzers.

AI-Generated-By: Opencode (Claude Opus 4.6)
