--- internal/repository/models/doc.go
@@ -1,8 +1,0 @@
-// Package models provides database entity models for the career domain.
-//
-// # Overview
-//
-// The models package contains:
-// - Database struct definitions for career entities
-// - ORM-compatible model types
-package models
