+++ tools/analyzers/noinlinecareer/analyzer_test.go
@@ -27,7 +27,7 @@ func TestFixturePackagesExcluded(t *testing.T) {
 	analysistest.Run(t, testdata, noinlinecareer.Analyzer, "fixtures")
 }
 
-func TestUnexportedTypesSkipped(t *testing.T) {
+func TestDomainCareerPackageExcluded(t *testing.T) {
 	testdata := analysistest.TestData()
-	analysistest.Run(t, testdata, noinlinecareer.Analyzer, "fake/career")
+	analysistest.Run(t, testdata, noinlinecareer.Analyzer, "fake/domain/career")
 }
