+++ tools/analyzers/noinlinecareer/testdata/src/fake/domain/career/types.go
