+++ tools/analyzers/noinlinecareer/testdata/src/fake/domain/career/types_test.go
@@ -1,7 +1,7 @@
 package career
 
 func helperWithExportedType() {
-	_ = Event{ID: "1"} // want `inline career\.Event\{\} in test file; use fixtures package instead`
+	_ = Event{ID: "1"}
 }
 
 func helperWithUnexportedType() {
