+++ tools/analyzers/noinlinecareer/testdata/src/nontest/nontest.go
@@ -1,6 +1,6 @@
 package nontest
 
-import "fake/career"
+import "fake/domain/career"
 
 func productionCodeWithCareerLiteral() career.Event {
 	return career.Event{ID: "1", Text: "allowed in non-test files"}
